+++ demo/src/lib.rs
@@ -4,10 +4,11 @@ mod plugins;
 mod windows;
 
 use std::collections::HashMap;
+use std::{cell::Cell, rc::Rc};
 
 use crate::plugins::ImagesPluginData;
 use egui::Context;
-use walkers::{HttpOptions, Map, MapMemory, Tiles, TilesManager};
+use walkers::{HttpOptions, Map, MapMemory, Position, Tiles, TilesManager};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum Provider {
@@ -94,6 +95,7 @@ pub struct MyApp {
     selected_provider: Provider,
     map_memory: MapMemory,
     images_plugin_data: ImagesPluginData,
+    last_click_position: Rc<Cell<Option<Position>>>,
 }
 
 impl MyApp {
@@ -108,6 +110,7 @@ impl MyApp {
             selected_provider: Provider::OpenStreetMap,
             map_memory: MapMemory::default(),
             images_plugin_data,
+            last_click_position: Rc::new(Cell::new(None)),
         }
     }
 }
@@ -139,7 +142,10 @@ impl eframe::App for MyApp {
                 let map = map
                     .with_plugin(plugins::places())
                     .with_plugin(plugins::images(&mut self.images_plugin_data))
-                    .with_plugin(plugins::CustomShapes {});
+                    .with_plugin(plugins::CustomShapes {})
+                    .with_plugin(plugins::ClickWatcher {
+                        last_click: self.last_click_position.clone(),
+                    });
 
                 // Draw the map widget.
                 ui.add(map);
@@ -150,6 +156,9 @@ impl eframe::App for MyApp {
 
                     zoom(ui, &mut self.map_memory);
                     go_to_my_position(ui, &mut self.map_memory);
+                    if let Some(position) = self.last_click_position.get() {
+                        show_my_position(ui, &position);
+                    }
                     controls(
                         ui,
                         &mut self.selected_provider,