+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+ * `TileManager` trait and demonstration of locally generated tiles.
+
 ## 0.15.0
 
  * `egui` updated to 0.24. This change requires Rust 1.72 or greater.
- * `TileManager` trait and demonstration of local-generated rainbowed tiles
 
 ## 0.14.0